move message handler somewhere more obvious
