+++ wampy/roles/callee.py
@@ -6,7 +6,7 @@ import logging
 import types
 from functools import partial
 
-from wampy.messages.handler import MessageHandler
+from wampy.message_handler import MessageHandler
 from wampy.peers.clients import Client
 
 logger = logging.getLogger(__name__)
